+++ lib/features/hero_information/domain/entities/trainer/trainer.dart
@@ -1,15 +1,15 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 
 import 'package:equatable/equatable.dart';
-import 'package:hero_flutter/features/hero_information/domain/entities/body_measurements_entity.dart';
-import 'package:hero_flutter/features/hero_information/domain/entities/training_entity.dart';
+import 'package:hero_flutter/features/hero_information/domain/entities/trainer/body_measurements_entity.dart';
+import 'package:hero_flutter/features/hero_information/domain/entities/trainer/training_entity.dart';
 
 class Trainer extends Equatable {
   final String? name;
-  final int? age;
+  final double? age;
   final String? gender;
-  final int? height;
-  final int? weight;
+  final double? height;
+  final double? weight;
   final String? goal;
   final TrainingEntity? training;
   final BodyMeasurementsEntity? bodyMeasurementsEntity;
